requirements.txt
crewai>=0.11.0
python-dotenv>=1.0.0
langchain>=0.1.0
chromadb>=0.4.22
tiktoken>=0.5.2
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2 